+++ packages/core/client/Alert.js
@@ -1,0 +1,2 @@
+
+Kepler.Alert = sAlert;