+++ packages/i18n/i18n/de.js
@@ -54,7 +54,8 @@ i18n.map('de', {
 	error_novalid: "not valid",
 	error_validchars: "letters, numbers, '-' '.' '_' min 3 max 30 chars",
 	error_taken: "username <i>{$1}</i> already been taken",
-	
+	error_nozoomdata: "Vergrößern, um Daten anzuzeigen",
+
 	btn_ok: "OK",
 	btn_cancel: "Abbrechen",
 	btn_checkin: "Einchecken",