+++ packages/i18n/i18n/es.js
@@ -53,7 +53,8 @@ i18n.map('es', {
 	error_novalid: "not valid",
 	error_validchars: "letters, numbers, '-' '.' '_' min 3 max 30 chars",
 	error_taken: "username <i>{$1}</i> already been taken",
-	
+	error_nozoomdata: "Acercar para mostrar datos",
+
 	btn_ok: "Ok",
 	btn_cancel: "Anular",
 	btn_checkin: "Adelante",