+++ packages/i18n/i18n/sv.js
@@ -54,6 +54,7 @@ i18n.map('sv', {
 	error_novalid: "not valid",
 	error_validchars: "letters, numbers, '-' '.' '_' min 3 max 30 chars",
 	error_taken: "username <i>{$1}</i> already been taken",
+	error_nozoomdata: "Zooma in för att visa data",
 	
 	btn_ok: "Ok",
 	btn_cancel: "Avbryt",