+++ packages/import/client/views/panels.js
@@ -3,7 +3,7 @@ Template.formImport.helpers({
 	importDone: function() {
 		var tmpl = Template.instance();
 		return function(ret) {
-			sAlert.info(ret)
+			K.Alert.info(ret)
 		}
 	}
 });